Registrer form update (#76)
* Dev (#74) Styret 26/27

* adding cicd build to dev branch

* Add styret 26/27

---------


* Move registration form to /registrer and add expected finish year

The registration form now lives at /registrer, replacing the easter-egg
page that hid a link to it. /den-ekte-registreringssiden redirects there
so existing links keep working.

Students are also asked which year they expect to finish their studies.
It is sent and validated as an ISO8601 string in study.expectedFinishYear.


---------
