+++ src/routes/registrer.svelte
@@ -7,17 +7,47 @@
 </script>
 
 <script>
-	import { onMount } from 'svelte';
-	import { fade } from 'svelte/transition';
+	import { goto } from '@sapper/app';
+	import Button from '../components/Button.svelte';
+	import api from '../utils/api';
 
-	const html = '<' + '!-- Hysj. Registreringsskjemaet er blitt flyttet hit: https://itemize.no/den-ekte-registreringssiden -->'; // issue https://github.com/sveltejs/prettier-plugin-svelte/issues/244
-
-	let showHelp = true;
-	onMount(() => setTimeout(() => (showHelp = true), 0 * 1000));
-	let expandHelp = false;
-	const toggleHelp = (e) => {
-		if (e.target.tagName === 'BUTTON') expandHelp = false;
-		else expandHelp = true;
+	const types = {
+		student: 'Student',
+		employee: 'Ansatt',
+		alumni: 'Alumni'
+	};
+	let user = {
+		fullName: '',
+		email: '',
+		data: {
+			type: Object.keys(types)[0],
+			displayName: '',
+			study: {
+				program: '',
+				year: 1
+			},
+			alumni: {
+				joinYear: new Date().getFullYear()
+			},
+			employee: {
+				title: ''
+			}
+		}
+	};
+	let expectedFinishYear = new Date().getFullYear();
+	// Studieslutt lagres som ISO8601 dato, 1. juni i det oppgitte året.
+	const yearToISO = (year) => (year && Number.isInteger(+year) ? new Date(Date.UTC(+year, 5, 1)).toISOString() : '');
+	let error = {
+		email: ''
+	};
+	let resp;
+	let register = async () => {
+		resp = null;
+		const payload = { ...user, data: { ...user.data, study: { ...user.data.study, expectedFinishYear: yearToISO(expectedFinishYear) } } };
+		resp = await api.registerUser(payload);
+		if (!resp.error) {
+			await goto('/registrert');
+		}
 	};
 </script>
 
@@ -32,87 +62,80 @@
 		Som medlem har du mulighet for å bli med på Discord-serveren vår, hvor all intern kommunikasjon foregår, samt mulighet for å bli med på prosjekter eller stemme ved
 		generalforsamling.
 	</p>
-	<br />
-	<p>
-		Hm... Det ser ut som det skulle ha vært et registreringsskjema her. Mulig noen har flyttet det? Kanskje det finnes noen spor gjemt i kildekoden på siden her? Mulig du kan finne
-		noe med å inspisere litt her på siden?
-	</p>
 
-	{#if showHelp}
-		<div class="help" class:expanded={expandHelp} transition:fade on:click={toggleHelp} title="Help?">
-			<div class="icon">?</div>
-
-			<div class="text">
-				De fleste nettlesere støtter å inspisere kildekoden til nettsider ved å høyre-klikke på siden og velg <code>Inspiser Element</code> eller lignende. Sjekk gjerne ut
-				<a target="_blank" rel="noopener noreferrer" href="https://ddg.co?q=how+to+inspect+element+in+BROWSER">her</a> for hjelp med din spesifike nettleser.
-				<br />
-				Hvis du er på mobil kan det være greit å prøve igjen når en har en PC tilgjengelig.
-				<button on:click={toggleHelp}>Lukk hintet</button>
-			</div>
-		</div>
+	{#if resp?.error}
+		<p class="error">{resp.error.replace(/.*failed custom validation because /, '') || 'Ups. Noe gikk galt :/'}</p>
 	{/if}
+	<div class="form">
+		<div class="cell">Fullt Navn:</div>
+		<div class="cell"><input type="text" bind:value={user.fullName} /></div>
+		<div class="comment">Fullt navn kan kun settes en gang.</div>
 
-	{@html html}
+		<div class="cell">E-post adresse:</div>
+		<div class="cell"><input type="text" bind:value={user.email} /></div>
+		{#if error.email}
+			<div class="comment error">{error.email}</div>
+		{/if}
+
+		<div class="cell">Visningsnavn:</div>
+		<div class="cell"><input type="text" bind:value={user.data.displayName} /></div>
+		<div class="comment">Ditt navn på nettsiden og i e-poster. Dette kan endres senere.</div>
+
+		<div class="cell">Medlemstype:</div>
+		<div class="cell">
+			<select bind:value={user.data.type}>
+				{#each Object.keys(types) as type}
+					<option value={type}>{types[type]}</option>
+				{/each}
+			</select>
+		</div>
+		{#if user.data.type == 'student'}
+			<div class="comment">Student ved NTNU Trondheim.</div>
+		{:else if user.data.type == 'alumni'}
+			<div class="comment">Tidligere Itemize medlem.</div>
+		{:else if user.data.type == 'employee'}
+			<div class="comment">Ansatt ved NTNU.</div>
+		{/if}
+
+		{#if user.data.type == 'student' || user.data.type == 'alumni'}
+			<div class="cell">Studieprogram:</div>
+			<div class="cell"><input type="text" bind:value={user.data.study.program} /></div>
+			{#if user.data.type == 'student'}
+				<div class="cell">Studieår:</div>
+				<div class="cell"><input type="number" min="1" max="100" bind:value={user.data.study.year} /></div>
+				<div class="comment">Nåværende progresjonsår i studiet. Vanligvis mellom 1 og 5.</div>
+				<div class="cell">Forventet ferdig år:</div>
+				<div class="cell"><input type="number" min={new Date().getFullYear()} max={new Date().getFullYear() + 10} bind:value={expectedFinishYear} /></div>
+				<div class="comment">Hvilket år regner du med å være ferdig med studiet ditt?</div>
+			{:else if user.data.type == 'alumni'}
+				<div class="cell">Medlems år:</div>
+				<div class="cell"><input type="number" min="2014" max={new Date().getFullYear()} bind:value={user.data.alumni.joinYear} /></div>
+				<div class="comment">Hvis du er tideligere alumni, hvilket år ble du først med i Itemize?</div>
+			{/if}
+		{:else if user.data.type == 'employee'}
+			<div class="cell">Title:</div>
+			<div class="cell"><input type="text" bind:value={user.data.employee.title} /></div>
+			<div class="comment">Fyll ut best beskrivende jobb tittel.</div>
+		{/if}
+	</div>
+	<Button fill big submit={register}>Register</Button>
 </main>
 
 <style>
-	main {
-		position: relative;
-		padding-bottom: 70px;
-	}
-	.help {
-		position: absolute;
-		padding: 0.25em;
-		min-width: 4em;
-		min-height: 4em;
-		bottom: -10px;
-		right: -10px;
-		margin: 2em;
-		background: #555;
-		border: 2px solid #888;
-		cursor: pointer;
-		font-size: 0.95em;
-		max-width: min(40ch, 50vw);
-		display: flex;
-		border-radius: 50%;
+	.form {
+		display: grid;
+		grid-template-columns: 50% 50%;
 		align-items: center;
-		justify-content: center;
 	}
-	.help div {
-		transition: transform 0.5s ease;
+	.form * {
+		margin: 0.4em 0;
 	}
-	.help .text {
-		width: 400px;
-		position: absolute;
-		bottom: 0;
-		opacity: 0;
-		pointer-events: none;
-		/* transform: scale(0.4); */
-	}
-	.help.expanded {
-		border-radius: 0;
-		cursor: auto;
-	}
-	.help.expanded .icon {
-		display: none;
-	}
-	.help.expanded .text {
-		position: relative;
-		opacity: 1;
-		pointer-events: all;
-		/* transform: scale(1); */
-		left: 0;
-	}
-	.icon {
-		text-align: center;
-		font-size: 2.5em;
-		line-height: 1;
-	}
-	.text {
-		padding: 10px;
-	}
-	button {
-		margin: 10px auto 0 auto;
+	.cell * {
 		width: 100%;
+	}
+	.comment {
+		grid-column: 1/3;
+		font-size: 0.8em;
+		margin: -1em 0 1em;
 	}
 </style>
